+++ src/main/java/gregtech/common/tileentities/machines/multi/nanochip/modules/MTESuperconductorSplitterModule.java
@@ -153,7 +153,7 @@ public class MTESuperconductorSplitterModule extends MTENanochipAssemblyModuleBa
 
         if (ticker % 20 == 0) {
             FluidStack fluidToBeDrained = MaterialLibAPI
-                .getFluidStack(Materials.SuperCoolant, FluidShapes.fluidLiquid, (int) (COOLANT_CONSUMED_PER_SEC));
+                .getFluidStack(Materials.SuperCoolant, FluidShapes.fluidLiquid, COOLANT_CONSUMED_PER_SEC);
             if (!drain(coolantInputHatch, fluidToBeDrained, true)) {
                 stopMachine(ShutDownReasonRegistry.outOfFluid(fluidToBeDrained));
                 return false;
