+++ src/main/java/gregtech/loaders/oreprocessing/ProcessingRawOre.java
@@ -140,9 +140,7 @@ public class ProcessingRawOre implements gregtech.api.interfaces.IOreRecipeRegis
             if (MaterialUtils.hasFlag(material, GTMaterialFlag.BLASTFURNACE_CALCITE_TRIPLE)) {
                 if (MaterialUtils.autoGenerateBlastFurnaceRecipes(material)) {
                     GTValues.RA.stdBuilder()
-                        .itemInputs(
-                            oreStack,
-                            MaterialLibAPI.getStack(Materials.Calcite, Shapes.dust, (int) (multiplier)))
+                        .itemInputs(oreStack, MaterialLibAPI.getStack(Materials.Calcite, Shapes.dust, multiplier))
                         .itemOutputs(
                             GTUtility.mul(3 * MaterialUtils.smeltingMultiplier(material), tSmeltInto),
                             GTOreDictUnificator.get(OrePrefixes.dust, Materials.DarkAsh, 1L))
@@ -152,9 +150,7 @@ public class ProcessingRawOre implements gregtech.api.interfaces.IOreRecipeRegis
                         .metadata(COIL_HEAT, 1500)
                         .addTo(blastFurnaceRecipes);
                     GTValues.RA.stdBuilder()
-                        .itemInputs(
-                            oreStack,
-                            MaterialLibAPI.getStack(Materials.Quicklime, Shapes.dust, (int) (multiplier)))
+                        .itemInputs(oreStack, MaterialLibAPI.getStack(Materials.Quicklime, Shapes.dust, multiplier))
                         .itemOutputs(
                             GTUtility.mul(3 * MaterialUtils.smeltingMultiplier(material), tSmeltInto),
                             GTOreDictUnificator.get(OrePrefixes.dust, Materials.DarkAsh, 1L))
@@ -167,9 +163,7 @@ public class ProcessingRawOre implements gregtech.api.interfaces.IOreRecipeRegis
             } else if (MaterialUtils.hasFlag(material, GTMaterialFlag.BLASTFURNACE_CALCITE_DOUBLE)) {
                 if (MaterialUtils.autoGenerateBlastFurnaceRecipes(material)) {
                     GTValues.RA.stdBuilder()
-                        .itemInputs(
-                            oreStack,
-                            MaterialLibAPI.getStack(Materials.Calcite, Shapes.dust, (int) (multiplier)))
+                        .itemInputs(oreStack, MaterialLibAPI.getStack(Materials.Calcite, Shapes.dust, multiplier))
                         .itemOutputs(
                             GTUtility.mul(2 * MaterialUtils.smeltingMultiplier(material), tSmeltInto),
                             GTOreDictUnificator.get(OrePrefixes.dust, Materials.DarkAsh, 1L))
@@ -179,9 +173,7 @@ public class ProcessingRawOre implements gregtech.api.interfaces.IOreRecipeRegis
                         .metadata(COIL_HEAT, 1500)
                         .addTo(blastFurnaceRecipes);
                     GTValues.RA.stdBuilder()
-                        .itemInputs(
-                            oreStack,
-                            MaterialLibAPI.getStack(Materials.Quicklime, Shapes.dust, (int) (multiplier)))
+                        .itemInputs(oreStack, MaterialLibAPI.getStack(Materials.Quicklime, Shapes.dust, multiplier))
                         .itemOutputs(
                             GTUtility.mul(2 * MaterialUtils.smeltingMultiplier(material), tSmeltInto),
                             GTOreDictUnificator.get(OrePrefixes.dust, Materials.DarkAsh, 1L))
